requirements.txt
pycryptodomex==3.11
bluepy
paho-mqtt
